In XIA xMap MCS mode, the detector electronics returns the data into a pair of buffers called 'buffer_a' and 'buffer_b' as 16-bit integers.  However, the arrays that are passed to MX are actually 32-bit arrays with the high order 16 bits always set to zero.  Modified the definitions of the 'buffer_a' and 'buffer_b' arrays accordingly.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@4293 42547f41-83fc-0310-9aef-9668e019ab9e

+++ modules/xia_handel_map/d_handel_mcs.c
@@ -616,7 +616,7 @@ mxd_handel_mcs_arm( MX_MCS *mcs )
 		handel_mcs->buffer_a = malloc( handel_mcs->buffer_length
 						* sizeof(long) );
 
-		if ( handel_mcs->buffer_a == (uint16_t *) NULL ) {
+		if ( handel_mcs->buffer_a == (uint32_t *) NULL ) {
 			return mx_error( MXE_OUT_OF_MEMORY, fname,
 			"Ran out of memory trying to allocate a "
 			"%lu element array of unsigned longs for "
@@ -628,7 +628,7 @@ mxd_handel_mcs_arm( MX_MCS *mcs )
 		handel_mcs->buffer_b = malloc( handel_mcs->buffer_length
 						* sizeof(long) );
 
-		if ( handel_mcs->buffer_b == (uint16_t *) NULL ) {
+		if ( handel_mcs->buffer_b == (uint32_t *) NULL ) {
 			return mx_error( MXE_OUT_OF_MEMORY, fname,
 			"Ran out of memory trying to allocate a "
 			"%lu element array of unsigned longs for "
